Add `ApplicationSystem` class
that contains all the device specific stuff (i.e. it represents the
actual hardware) and have the `Application` class handle only the SiLA 2
servers (i.e. the software representation of the system). That way, the
Feature implementation classes can access the `ApplicationSystem`
instance (which is already fully constructed at that point) in their
constructors. Before, when the `Application` class handled everything,
the Feature implementation classes couldn't access the `Application`
instance as it wasn't fully constructed yet because it was in the middle
of creating the SiLA 2 servers.

Change-Id: I2830a4eefbbddb12ae0bd7f624314fa1d2a91891