Update `__init__.py`
to setup the environment for python and ctypes to find the qmixsdk
package as well as the CETONI SDK shared libraries. Reruns the same
command (i.e. either `python -m sila_cetoni ...` or `sila-cetoni ...`)
with the correct environment. This finally removes the need for an
additional wrapper shell script for Linux systems.