Add `--scan` option to force scanning for devices
like serial devices (e.g. Sartorius balance) and change the default
behaviour to not scan for devices.

On RevPi systems the automatic scanning of serial ports leads to an
unrecoverable state of the `piControl` process due to opening the
`/dev/ttyAMA0` device. To prevent this we don't scan for devices
automatically from now on. In the future there will be a way to
statically define the type and number of devices to use (either via a
configuration file or via CLI options) since this is fixed anyway most
of the time.