Add config.json file concept
changing how to define the devices for which SiLA servers should be
started.

A config.json file is a JSON description of the whole application
system. It contains a description of the 3rd-party devices for which to
create servers as well as an optional path to a CETONI device
configuration folder which should also be used to create the
corresponding servers. In addition, this file can be used to define
persistent values for the CLI parameters. The CLI parameters can still
be used to override the values from the config file, however.

Due to this change the internal structure of `Application`,
`ApplicationSystem`, `Device`, etc. has been changed, as well.
Internally, we now distinguish between a `CetoniApplicationSystem` and
the overall `ApplicationSystem`. The latter optionally contains the
former if specified in the config file. There is also a distinction
between `CetoniDevice`s and `ThirdPartyDevice`s. However, the goal is
that you can use a `ThirdPartyDevice` in place of a `CetoniDevice` and
vice versa when it comes to implementing features.

The main application learned a new option `--log-file-dir` which given a
directory path will put a log file of the current execution into that
directory.
Also, the main application now has a mandatory argument: the path to the
config.json file (which doesn't have to be called 'config.json', in fact
it should have a more descriptive way so that multiple configurations
can coexist in the same folder and it's as easy as changing one argument
when calling `sila-cetoni` or `python -m sila_cetoni.application`).