Fix that "not found" exceptions weren't caught
in `ApplicationSystem.start`. Since these exceptions can happen in
`start` as well as in the `__create_..._devices` functions there is the
possibility that the device that caused the exception should be set to
simulated. Therefore this is now done by a separate function that either
sets the device simulated or raises the "not found" exception that was
caught.