Add `monitor_traffic` decorator to `CetoniApplicationSystem`
that can be used to decorate all feature implementation classes that
should be monitored for traffic. If none of the monitored class's
methods is called within a certain time (currently 10 minutes) and the
device is battery powered then the device is (gracefully) shut down. If
it is not battery powered then it won't be shut down.

In case of our MobDos this means that it will only be shut down if it is
**solely** powered by a battery. If the external power is connected then
it is not shut down since there is no reason to save power in this case.