Fix segmentation fault for shutdown over `Shutdown` Command
because `ApplicationSystem.shutdown` (and subsequently `stop` which
calls `qmixbus.Bus.stop()`/`.close()` were not run in the main thread.
This is now fixed by detecting that the call is not executed in the main
thread and moving the call to the main thread where the `Application` is
polling a tasks queue while waiting for the SiLA Server(s) to exit.