Update `ServerConfiguration` to behave like `ConfigParser`
I.e. add methods for subscript (`[ ]`) access, `len`, `iter` and `in`.
This has the advantage that add-on packages can dynamically add new
sections and store their config values as they see fit, without having
to modify `ServerConfiguration` each time. The add-on packages have to
take care to create the sections in case they're not present yet.