Fix that only specifying the log file dir in the JSON config
didn't start logging to the file. To have the log messages from parsing
the `"devices"` section in the JSON file also appear in the log file if
the log file dir is only specified in the JSON config,
`ApplicationConfiguration` is changed so that `Application.run` will
parse the `"devices"` section and `ApplicationConfiguration.__init__`
will only parse the top-level keys. This still means that some log
messages will not be written to the log file if the log file dir is only
given in the JSON config (e.g. messages about importing available add-on
packages and their JSON schemas) but this is to be expected. If this is
not desired then the CLI option must be used.