Update `CetoniApplicationSystem.__monitor_events`
When the EPOS sends the "DC link under voltage" event it might be that
the battery is not yet removed but rather the voltage is too low for the
EPOS to function (i.e. < 24V but > 22.5V, which is the absolute
minimum). In this case the `seconds_stopped` would already be counted
even though the battery was still plugged in and a battery replacement
could already be initiated. To fix this, `seconds_stopped` will now only
start being counted if the battery is actually removed. Additionally,
the "DC link under voltage" event is now translated to and error that
is posted to the ErrorProvider of the `MobDosServer` to inform the user
about the fact that the pump drive can no longer be moved.

Fixes #1464