Add support for the same server to be run on different IP addresses
but with the same UUID

To achieve this `ServerConfiguration` gets a new method called
`ssl_certificate_for_ip(ip)` which checks if that given IP address is
already present in the SANs of the certificate. If not that IP is added
and the new certificate is stored in the config file as well as returned
to the caller (`Application` in this case) who can then use that
certificate to start the server. This alleviates the process of having
to regenerate the certificate manually each time the IP address of the
host changes.

Implements #1479