Update `__init__.py`
Attempt to import PySide2 before importing qmixsdk. This has to be done
in case the Qt Python bindings should be used. The bindings need to be
imported before qmixsdk in order for the Qt bindings to use their DLLs
and not the ones from qmixsdk. qmixsdk can use the DLLs from PySide2 but
not the other way around.