Change what happens when the `server_base_port` is changed
If a server had been started with a certain base port before but now the
base port is overwritten explicitly by the CLI option, this new port is
now respected and the port from the previous run will be overwritten
with a port number calculated from the new base port. This now allows to
change the port of a server which had been run before on a different
port (e.g. if that port is no longer available to use) without having to
delete or manually edit the server config file.