Fix that there was no way to force a change of the server port
if the server config file did not contain the server base port that was
used when the server was first started. In this case it wasn't possible
to compare the current and the previous server base port to detect
wether it changed. This commit changes the behavior of the
`-p`/`--server-base-port` CLI option to always force a change of the
server base port (and thus the server port of all servers) even if the
server config file does not contain the previous server base port.